+++ NEWS.md
@@ -1,3 +1,8 @@
+- 2011-12-17
+    - web-socket-js now uses MozWebSocket when available. i.e. When you load
+      web_socket.js, WebSocket is defined as alias of MozWebSocket when
+      available.
+
 - 2011-09-18
     - web-socket-js now speaks WebSocket version hybi-10. Old versions spoke
       hixie-76. If you really need web-socket-js which speaks hixie-76, you can